Fixed tests; split options type module
